Move helper classes to helpers/ folder
