+++ lib/centra/helpers/loggable.rb
