+++ lib/centra/helpers/null_logger.rb
